refactor: worker 공통 로직 shared 모듈로 추출
